+++ common/src/main/java/com/yanny/ali/plugin/mods/hybrid_aquatic/Plugin.java
@@ -1,0 +1,19 @@
+package com.yanny.ali.plugin.mods.hybrid_aquatic;
+
+import com.yanny.ali.api.AliEntrypoint;
+import com.yanny.ali.api.IPlugin;
+import com.yanny.ali.api.IServerRegistry;
+import com.yanny.ali.plugin.mods.PluginUtils;
+
+@AliEntrypoint
+public class Plugin implements IPlugin {
+    @Override
+    public String getModId() {
+        return "hybrid-aquatic";
+    }
+
+    @Override
+    public void registerServer(IServerRegistry registry) {
+        PluginUtils.registerEntry(registry, MessageInABottleItemEntry.class);
+    }
+}