+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added compatibility for Villagers Plus
 - Added compatibility for Cultural Delights
 - Added compatibility for Sawmill
+- Added compatibility for Hybrid Aquatic
 
 ## [1.3.3]